[Auto Generated] 0.1.28

+++ fdk_client/application/models/ShipmentTotalDetails.py
@@ -16,9 +16,9 @@ class ShipmentTotalDetails(BaseSchema):
     # Order swagger.json
 
     
-    sizes = fields.Float(required=False)
+    sizes = fields.Int(required=False)
     
-    pieces = fields.Float(required=False)
+    pieces = fields.Int(required=False)
     
     total_price = fields.Float(required=False)
